Add an application configuration file
Fixes 5G-MAG/rt-5gms-application-server#5

Adds an application configuration file (example in
docs/example-application-server.conf) which contains the directory to place the
cache into (default: /var/cache/rt-5gms/as/cache). If set to the empty string
will disable caching. This configuration also allows the HTTP and HTTPS
listening port numbers to be set in the configuration to aid in development
(i.e. http_port and https_port can be set above 1024 for running as a normal
user). The configuration also provides a section for nginx specific
configuration values and so provides a template for adding configuration for
other backend web servers/proxies.