Add Certificate handling for HTTPS distrubution.
Closes 5G-MAG/rt-5gms-application-server#9.

This will use an optional second command line parameter to specify a Certificates JSON file. This file is a simple map of certificate ID to certificate file path. If a relative path is used in the mapping then it will be treated as relative to the Certificates JSON file. This allows certificates to be easily located with Certificates JSON file.

If the ContentHostingConfiguration references a certificate, by its ID, in a distributionConfiguration, then an HTTPS distribution will be created using the certificate and key in the file pointed to by the Certificates JSON file. Multiple distribution points can be provided to use different server certificates or to provide both HTTP and HTTPS distribution points for the same ingest material.

Instructions to run the examples have been provided in the `docs/README.md` file.